udp-native folder stucture refactored
