+++ udp-native/src/main/java/org/ballerinalang/stdlib/udp/ReadPendingCallback.java
@@ -16,11 +16,10 @@
  * under the License.
  */
 
-package org.ballerinalang.stdlib.socket.tcp;
+package org.ballerinalang.stdlib.udp;
 
 import io.ballerina.runtime.api.Future;
 import io.ballerina.runtime.api.values.BError;
-import org.ballerinalang.stdlib.socket.SocketConstants;
 
 import java.nio.ByteBuffer;
 import java.util.Timer;
