+++ udp-native/src/main/java/org/ballerinalang/stdlib/udp/ReadReadySocketMap.java
@@ -16,13 +16,15 @@
  * under the License.
  */
 
-package org.ballerinalang.stdlib.socket.tcp;
+package org.ballerinalang.stdlib.udp;
+
+import org.ballerinalang.stdlib.udp.endpoint.ClientActions;
 
 import java.util.concurrent.ConcurrentHashMap;
 
 /**
  * This map will hold all the read ready sockets until it is read through the
- * {@link org.ballerinalang.stdlib.socket.endpoint.udp.ClientActions} read action.
+ * {@link ClientActions} read action.
  *
  * @since 0.995.0
  */
