+++ udp-native/src/main/java/org/ballerinalang/stdlib/udp/SocketThreadFactory.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.ballerinalang.stdlib.socket;
+package org.ballerinalang.stdlib.udp;
 
 import java.util.concurrent.ThreadFactory;
 import java.util.concurrent.atomic.AtomicInteger;
