+++ udp-native/src/main/resources/META-INF/services/org.ballerinalang.compiler.plugins.CompilerPlugin
@@ -16,4 +16,4 @@
 # under the License.
 #
 
-org.ballerinalang.stdlib.socket.compiler.SocketCompilerPlugin
+org.ballerinalang.stdlib.udp.compiler.SocketCompilerPlugin
