+++ udp-test-utils/src/main/java/module-info.java
@@ -20,5 +20,5 @@ module io.ballerina.stdlib.socket.testutils {
     requires io.ballerina.runtime;
     requires org.slf4j;
     requires io.ballerina.stdlib.socket;
-    exports org.ballerinalang.stdlib.socket.testutils;
+    exports org.ballerinalang.stdlib.udp.testutils;
 }
